CFHydrostatics

Documentation for CFHydrostatics.

CFHydrostatics.PressureCoordinate

CFHydrostatics.SigmaCoordinate

CFHydrostatics.VerticalCoordinate

CFHydrostatics.diagnostics

CFHydrostatics.initial_HPE

CFHydrostatics.pressure_level

CFHydrostatics.PressureCoordinate — Type

abstract type PressureCoordinate{N} <: VerticalCoordinate{N} end

Parent type for a pressure-based vertical coordinate. Children types should specialize pressure_level. See also VerticalCoordinate.

CFHydrostatics.SigmaCoordinate — Type

sigma = SigmaCoordinate(N, ptop) <: PressureCoordinate{N}

Pressure based sigma-coordinate for N levels with top pressure ptop. Pressure levels are linear in vertical coordinate k : k/N = (ps-p)/(ps-ptop) where k ranges from 0 (ground) to N (model top).

CFHydrostatics.VerticalCoordinate — Type

abstract type VerticalCoordinate{N} end

Parent type for generalized vertical coordinates ranging from 0 to N. See also PressureCoordinate.

CFHydrostatics.diagnostics — Method

diags = diagnostics(model::HPE)

Return diags::CookBook containing recipes to compute standard diagnostics. This object is to be used as follows:

session = open(diags; model, state)
temp = session.temperature
...

with state the current state, obtained for instance from initial_HPE.

CFHydrostatics.initial_HPE — Method

state = initial_HPE(case, model)

Return an initial state for the hydrostatic model. case is expected to be a function with two methods:

surface_pressure, surface_geopotential = case(longitude, latitude)
geopotential, ulon, ulat = case(longitude, latitude, pressure)

If the hydrostatic model is for a binary fluid, the second method should return instead:

geopotential, ulon, ulat, q = case(longitude, latitude, pressure)

CFHydrostatics.pressure_level — Function

p = pressure_level(k, ps, vcoord::PressureCoordinate{N})

Given surface pressure ps, returns the value p of pressure corresponding to level k/2 for vertical coordinate vcoord. This means that so-called full levels correspond to odd values k=1,2...2N-1 while interfaces between full levels (so-called half-levels) correspond to even values k=0,2...2N